+++ lib/src/components/effects/beauty_effect_button.dart
@@ -3,18 +3,18 @@ import 'package:flutter/material.dart';
 
 // Package imports:
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:zego_uikit/zego_uikit.dart';
 
 // Project imports:
 import 'package:zego_uikit_prebuilt_live_streaming/src/components/effects/beauty_effect_sheet.dart';
 import 'package:zego_uikit_prebuilt_live_streaming/src/internal/internal.dart';
-import 'package:zego_uikit_prebuilt_live_streaming/src/live_streaming_translation.dart';
+import 'package:zego_uikit_prebuilt_live_streaming/zego_uikit_prebuilt_live_streaming.dart';
 
 class ZegoBeautyEffectButton extends StatefulWidget {
   const ZegoBeautyEffectButton({
     Key? key,
     required this.beautyEffects,
     required this.translationText,
+    required this.rootNavigator,
     this.iconSize,
     this.buttonSize,
     this.icon,
@@ -25,6 +25,7 @@ class ZegoBeautyEffectButton extends StatefulWidget {
   final ButtonIcon? icon;
   final List<BeautyEffectType> beautyEffects;
   final ZegoTranslationText translationText;
+  final bool rootNavigator;
 
   @override
   State<ZegoBeautyEffectButton> createState() => _ZegoBeautyEffectButtonState();
@@ -40,6 +41,7 @@ class _ZegoBeautyEffectButtonState extends State<ZegoBeautyEffectButton> {
         showBeautyEffectSheet(
           context,
           translationText: widget.translationText,
+          rootNavigator: widget.rootNavigator,
           beautyEffects: widget.beautyEffects,
         );
       },