+++ lib/src/components/effects/sound_effect_button.dart
@@ -3,12 +3,11 @@ import 'package:flutter/material.dart';
 
 // Package imports:
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:zego_uikit/zego_uikit.dart';
 
 // Project imports:
 import 'package:zego_uikit_prebuilt_live_streaming/src/components/effects/sound_effect_sheet.dart';
 import 'package:zego_uikit_prebuilt_live_streaming/src/internal/internal.dart';
-import 'package:zego_uikit_prebuilt_live_streaming/src/live_streaming_translation.dart';
+import 'package:zego_uikit_prebuilt_live_streaming/zego_uikit_prebuilt_live_streaming.dart';
 
 class ZegoSoundEffectButton extends StatefulWidget {
   final List<VoiceChangerType> voiceChangeEffect;
@@ -18,10 +17,12 @@ class ZegoSoundEffectButton extends StatefulWidget {
   final Size? buttonSize;
   final ButtonIcon? icon;
   final ZegoTranslationText translationText;
+  final bool rootNavigator;
 
   const ZegoSoundEffectButton({
     Key? key,
     required this.translationText,
+    required this.rootNavigator,
     required this.voiceChangeEffect,
     required this.reverbEffect,
     this.iconSize,
@@ -46,6 +47,7 @@ class _ZegoSoundEffectButtonState extends State<ZegoSoundEffectButton> {
         showSoundEffectSheet(
           context,
           translationText: widget.translationText,
+          rootNavigator: widget.rootNavigator,
           voiceChangeEffect: widget.voiceChangeEffect,
           voiceChangerSelectedIDNotifier: voiceChangerSelectedIDNotifier,
           reverbEffect: widget.reverbEffect,